Add new library for RPC server module concept
All RPC servers share a common pattern. We use this generalization to
minimize the chance of making errors, and to minimize inconsistencies.

A server has to register three callbacks. On one callback, new messages
will arrive. The argument includes the RPC state, which includes itself
a pointer to some data which can be managed by the concrete server.
Another callback is used to initialize the RPC state by the server
implementation. However, channel related data is managed by the library
itself. The third callback is unused for now. When the IPC is fully
working, and process termination is correctly implemented, we can use
this callback to free server-specific data in the RPC state.